Deleted "testing" folder. Moved .java files into "towerPackage"
