+++ src/towerpackage/MovingTest.java
@@ -1,4 +1,4 @@
-package testing;
+package towerpackage;
 import java.awt.BorderLayout;
 import java.awt.EventQueue;
 import java.awt.Graphics;
